Extract attestation production and sending logic into a separate strategy to support feature toggling.
