+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/duties/ProductionResult.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2021 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.validator.client.duties;
+
+import java.util.Optional;
+import org.apache.tuweni.bytes.Bytes32;
+import tech.pegasys.teku.bls.BLSPublicKey;
+
+public class ProductionResult<T> {
+
+  private final BLSPublicKey validatorPublicKey;
+  private final DutyResult result;
+  private final Optional<T> message;
+
+  public ProductionResult(
+      final BLSPublicKey validatorPublicKey, final Bytes32 blockRoot, final T message) {
+    this.validatorPublicKey = validatorPublicKey;
+    this.result = DutyResult.success(blockRoot);
+    this.message = Optional.of(message);
+  }
+
+  public ProductionResult(final BLSPublicKey validatorPublicKey, final DutyResult result) {
+    this.validatorPublicKey = validatorPublicKey;
+    this.result = result;
+    this.message = Optional.empty();
+  }
+
+  public BLSPublicKey getValidatorPublicKey() {
+    return validatorPublicKey;
+  }
+
+  public DutyResult getResult() {
+    return result;
+  }
+
+  public boolean producedMessage() {
+    return message.isPresent();
+  }
+
+  public boolean failedToProduceMessage() {
+    return message.isEmpty();
+  }
+
+  public Optional<T> getMessage() {
+    return message;
+  }
+}
