+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/duties/attestations/AttestationProductionStrategy.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2021 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.validator.client.duties.attestations;
+
+import java.util.Map;
+import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.spec.datastructures.state.ForkInfo;
+import tech.pegasys.teku.validator.client.duties.DutyResult;
+
+public interface AttestationProductionStrategy {
+  SafeFuture<DutyResult> produceAttestations(
+      UInt64 slot, ForkInfo forkInfo, Map<Integer, ScheduledCommittee> validatorsByCommitteeIndex);
+}
