+++ validator/client/src/test/java/tech/pegasys/teku/validator/client/duties/AttestationProductionDutyTest.java
@@ -50,6 +50,8 @@ import tech.pegasys.teku.validator.api.SubmitDataError;
 import tech.pegasys.teku.validator.api.ValidatorApiChannel;
 import tech.pegasys.teku.validator.client.ForkProvider;
 import tech.pegasys.teku.validator.client.Validator;
+import tech.pegasys.teku.validator.client.duties.attestations.AttestationProductionDuty;
+import tech.pegasys.teku.validator.client.duties.attestations.IndividualAttestationProductionStrategy;
 
 class AttestationProductionDutyTest {
 
@@ -64,7 +66,8 @@ class AttestationProductionDutyTest {
   private final ValidatorLogger validatorLogger = mock(ValidatorLogger.class);
 
   private final AttestationProductionDuty duty =
-      new AttestationProductionDuty(SLOT, forkProvider, validatorApiChannel);
+      new AttestationProductionDuty(
+          SLOT, forkProvider, new IndividualAttestationProductionStrategy(validatorApiChannel));
 
   @BeforeEach
   public void setUp() {
